+++ ShipEngine/Models/Dto/Common/Enums/AddressResidentialIndicator.cs
@@ -1,0 +1,16 @@
+using System.Runtime.Serialization;
+
+namespace ShipEngineSDK.Common.Enums
+{
+    public enum AddressResidentialIndicator
+    {
+        [EnumMember(Value = "unknown")]
+        Unknown,
+
+        [EnumMember(Value = "yes")]
+        Yes,
+
+        [EnumMember(Value = "no")]
+        No,
+    }
+}
