+++ ShipEngine/Models/Dto/Common/Enums/BillToParty.cs
@@ -1,0 +1,12 @@
+using System.Runtime.Serialization;
+
+namespace ShipEngineSDK.Common.Enums
+{
+    public enum BillToParty
+    {
+        Recipient,
+
+        [EnumMember(Value = "third_party")]
+        ThirdParty
+    }
+}
