+++ ShipEngine/Models/Dto/Common/Enums/Currency.cs
@@ -1,6 +1,3 @@
-using Newtonsoft.Json;
-using Newtonsoft.Json.Converters;
-
 namespace ShipEngineSDK.Common.Enums
 {
     public enum Currency
