+++ ShipEngine/Models/Dto/Common/Enums/IdentifierType.cs
@@ -1,0 +1,14 @@
+namespace ShipEngineSDK.Common.Enums
+{
+    public enum IdentifierType
+    {
+        Vat,
+        Eori,
+        SSN,
+        Ein,
+        Tin,
+        Ioss,
+        Pan,
+        Voec
+    }
+}
