+++ ShipEngine/Models/Dto/Common/Enums/InsuranceProvider.cs
@@ -1,0 +1,14 @@
+using System.Runtime.Serialization;
+
+namespace ShipEngineSDK.Common.Enums
+{
+    public enum InsuranceProvider
+    {
+        None,
+        Shipsurance,
+        Carrier,
+
+        [EnumMember(Value = "third_party")]
+        ThirdParty
+    }
+}
