+++ ShipEngine/Models/Dto/Common/Enums/OriginType.cs
@@ -1,0 +1,12 @@
+using System.Runtime.Serialization;
+
+namespace ShipEngineSDK.Common.Enums
+{
+    public enum OriginType
+    {
+        Pickup,
+
+        [EnumMember(Value = "drop_off")]
+        DropOff
+    }
+}
