+++ ShipEngine/Models/Dto/Common/Enums/PaymentType.cs
@@ -1,0 +1,14 @@
+using System.Runtime.Serialization;
+
+namespace ShipEngineSDK.Common.Enums
+{
+    public enum PaymentType
+    {
+        Any,
+        Cash,
+
+        [EnumMember(Value = "cash_equivalent")]
+        CashEquivalent,
+        None
+    }
+}
