+++ ShipEngine/Models/Dto/Common/Enums/TaxableEntityType.cs
@@ -1,0 +1,9 @@
+namespace ShipEngineSDK.Common.Enums
+{
+    public enum TaxableEntityType
+    {
+        Shipper,
+        Recipient
+    }
+
+}
