+++ ShipEngine/Models/Dto/Common/MonetaryValue.cs
@@ -1,0 +1,12 @@
+
+using ShipEngineSDK.Common.Enums;
+
+namespace ShipEngineSDK.Common
+{
+    public class MonetaryValue
+    {
+        public Currency? Currency { get; set; }
+
+        public double? Amount { get; set; }
+    }
+}
