+++ ShipEngine/Models/Dto/CreateLabelFromShipmentDetails/Params.cs
@@ -1,3 +1,4 @@
+using ShipEngineSDK.Common;
 using ShipEngineSDK.Common.Enums;
 using System.Collections.Generic;
 
@@ -189,179 +190,6 @@ namespace ShipEngineSDK.CreateLabelFromShipmentDetails.Params
         public string? ExternalPackageId { get; set; }
     }
 
-    public class LabelMessages
-    {
-
-        /// <summary>
-        /// The first line of the custom label message.
-        /// Some carriers may prefix this line with something like "REF", "Reference", "Trx Ref No.", etc.
-        /// </summary>
-        public string? Reference1 { get; set; }
-
-        /// <summary>
-        /// The second line of the custom label message.
-        /// Some carriers may prefix this line with something like "INV", "Reference 2", "Trx Ref No.", etc.
-        /// </summary>
-        public string? Reference2 { get; set; }
-
-        /// <summary>
-        /// The third line of the custom label message.
-        /// Some carriers may prefix this line with something like "INV", "Reference 2", "Trx Ref No.", etc.
-        /// </summary>
-        public string? Reference3 { get; set; }
-    }
-
-    public class Dimensions
-    {
-        /// <summary>
-        /// The dimension units that are supported by ShipEngine.
-        /// </summary>
-        public DimensionUnit? Unit { get; set; }
-
-        /// <summary>
-        /// The length of the package, in the specified unit
-        /// </summary>
-        public double? Length { get; set; }
-
-        /// <summary>
-        /// The width of the package, in the specified unit
-        /// </summary>
-        public double? Width { get; set; }
-
-        /// <summary>
-        /// The length of the package, in the specified unit
-        /// </summary>
-        public double? Height { get; set; }
-    }
-
-    public class ShipmentItem
-    {
-
-        /// <summary>
-        /// Item name
-        /// </summary>
-        public string? Name { get; set; }
-
-        /// <summary>
-        /// Sales order id
-        /// </summary>
-        public string? SalesOrderId { get; set; }
-
-        /// <summary>
-        /// Sales order item id
-        /// </summary>
-        public string? SalesOrderItemId { get; set; }
-
-        /// <summary>
-        /// The quantity of this item included in a shipment
-        /// </summary>
-        public double? Quantity { get; set; }
-
-        /// <summary>
-        /// The item stock keeping unit
-        /// </summary>
-        public string? Sku { get; set; }
-
-        /// <summary>
-        /// External order id
-        /// </summary>
-        public string? ExternalOrderId { get; set; }
-
-        /// <summary>
-        /// External order item id
-        /// </summary>
-        public string? ExternalOrderItemId { get; set; }
-
-        /// <summary>
-        /// Amazon Standard Identification Number
-        /// </summary>
-        public string? Asin { get; set; }
-
-        /// <summary>
-        /// The order sources that are supported by ShipEngine
-        /// </summary>
-        public OrderSourceCode? OrderSourceCode { get; set; }
-    }
-    public class TaxIdentifier
-    {
-        /// <summary>
-        /// The taxable entity type for this tax item. Valid values include the following
-        /// </summary>
-        public TaxableEntityType? TaxableEntityType { get; set; }
-
-        /// <summary>
-        /// Determines how FedEx will pickup your packages
-        /// </summary>
-        public IdentifierType? IdentifierType { get; set; }
-
-        /// <summary>
-        /// The authority that issued this tax. This must be a valid 2 character ISO 3166 Alpha 2 country code.
-        /// </summary>
-        public string? IssuingAuthority { get; set; }
-
-        /// <summary>
-        /// The value of the identifier
-        /// </summary>
-        public string? Value { get; set; }
-    }
-    public class Address
-    {
-        /// <summary>
-        /// The first line of the address.
-        /// </summary>
-        public string? AddressLine1 { get; set; }
-
-        /// <summary>
-        /// The second line of the address.
-        /// </summary>
-        public string? AddressLine2 { get; set; }
-
-        /// <summary>
-        /// The third line of the address.
-        /// </summary>
-        public string? AddressLine3 { get; set; }
-
-        /// <summary>
-        /// The <see href="https://en.wikipedia.org/wiki/List_of_ISO_3166_country_codes">ISO 3166 country code</see>
-        /// </summary>
-        public Country? CountryCode { get; set; }
-
-        /// <summary>
-        /// The name of the sender or recipient at the address, if applicable.
-        /// </summary>
-        public string? Name { get; set; }
-
-        /// <summary>
-        /// The company name, if this is a business address.
-        /// </summary>
-        public string? CompanyName { get; set; }
-
-        /// <summary>
-        /// The phone number associated with this address, if any.
-        /// </summary>
-        public string? Phone { get; set; }
-
-        /// <summary>
-        /// The city or locality
-        /// </summary>
-        public string? CityLocality { get; set; }
-
-        /// <summary>
-        /// The state or province
-        /// </summary>
-        public string? StateProvince { get; set; }
-
-        /// <summary>
-        /// The postal code
-        /// </summary>
-        public string? PostalCode { get; set; }
-
-        /// <summary>
-        /// Indicates whether the address is residential or commercial, if known.
-        /// </summary>
-        public AddressResidentialIndicator? AddressResidentialIndicator { get; set; }
-    }
-
     public class InternationalShipmentOptions
     {
         /// <summary>
@@ -417,141 +245,5 @@ namespace ShipEngineSDK.CreateLabelFromShipmentDetails.Params
         /// Description of the Custom Item's SKU
         /// </summary>
         public string? SkuDescription { get; set; }
-    }
-    public class MonetaryValue
-    {
-        /// <summary>
-        /// The currencies that are supported by ShipEngine.
-        /// </summary>
-        public Currency? Currency { get; set; }
-
-        /// <summary>
-        /// The monetary amount, in the specified currency.
-        /// </summary>
-        public double? Amount { get; set; }
-    }
-    public class AdvancedShipmentOptions
-    {
-        /// <summary>
-        /// This field is used to bill shipping costs to a third party.
-        /// This field must be used in conjunction with the bill_to_country_code, bill_to_party, and bill_to_postal_code fields.
-        /// </summary>
-        public string? BillToAccount { get; set; }
-
-        /// <summary>
-        /// The two-letter ISO 3166-1 country code of the third-party that is responsible for shipping costs.
-        /// </summary>
-        public string? BillToCountryCode { get; set; }
-
-        /// <summary>
-        /// Indicates whether to bill shipping costs to the recipient or to a third-party.
-        /// When billing to a third-party, the bill_to_account, bill_to_country_code, and bill_to_postal_code fields must also be set.
-        /// </summary>
-        public BillToParty? BillToParty { get; set; }
-
-        /// <summary>
-        /// The postal code of the third-party that is responsible for shipping costs.
-        /// </summary>
-        public string? BillToPostalCode { get; set; }
-
-        /// <summary>
-        /// Indicates that the shipment contains alcohol.
-        /// </summary>
-        public bool? ContainsAlcohol { get; set; }
-
-        /// <summary>
-        /// Indicates that the shipper is paying the international delivery duties for this shipment.
-        /// This option is supported by UPS, FedEx, and DHL Express.
-        /// </summary>
-        public bool? DeliveredDutyPaid { get; set; }
-
-        /// <summary>
-        /// Indicates if the shipment contain dry ice
-        /// </summary>
-        public bool? DryIce { get; set; }
-
-        /// <summary>
-        /// The weight of the dry ice in the shipment
-        /// </summary>
-        public Weight? DryIceWeight { get; set; }
-
-        /// <summary>
-        /// Indicates that the package cannot be processed automatically because it is too large or irregularly shaped.
-        /// This is primarily for USPS shipments. See Section 1.2 of the USPS parcel standards for details.
-        /// </summary>
-        public bool? NonMachinable { get; set; }
-
-        /// <summary>
-        /// Enables Saturday delivery, if supported by the carrier.
-        /// </summary>
-        public bool? SaturdayDelivery { get; set; }
-
-        /// <summary>
-        /// Whether to use UPS Ground Freight pricing. If enabled, then a freight_class must also be specified.
-        /// </summary>
-        public bool? UseUpsGroundFreightPricing { get; set; }
-
-        /// <summary>
-        /// The National Motor Freight Traffic Association freight class, such as "77.5", "110", or "250".
-        /// </summary>
-        public string? FreightClass { get; set; }
-
-        /// <summary>
-        /// An arbitrary field that can be used to store information about the shipment.
-        /// </summary>
-        public string? CustomField1 { get; set; }
-
-        /// <summary>
-        /// An arbitrary field that can be used to store information about the shipment.
-        /// </summary>
-        public string? CustomField2 { get; set; }
-
-        /// <summary>
-        /// An arbitrary field that can be used to store information about the shipment.
-        /// </summary>
-        public string? CustomField3 { get; set; }
-
-        /// <summary>
-        /// Indicates if the package will be picked up or dropped off by the carrier
-        /// </summary>
-        public OriginType? OriginType { get; set; }
-
-        public bool? ShipperRelease { get; set; }
-
-        /// <summary>
-        /// Defer payment until package is delivered, instead of when it is ordered.
-        /// </summary>
-        public CollectOnDelivery? CollectOnDelivery { get; set; }
-    }
-    public class Weight
-    {
-        /// <summary>
-        /// The weight, in the specified unit
-        /// </summary>
-        public double? Value { get; set; }
-
-        /// <summary>
-        /// The possible weight unit values
-        /// </summary>
-        public WeightUnit? Unit { get; set; }
-    }
-    public class CollectOnDelivery
-    {
-        /// <summary>
-        /// Types of payment that are supported
-        /// </summary>
-        public PaymentType? PaymentType { get; set; }
-
-        public PaymentAmount? PaymentAmount { get; set; }
-    }
-
-    public class PaymentAmount
-    {
-        /// <summary>
-        /// The currencies that are supported by ShipEngine.
-        /// </summary>
-        public Currency? Currency { get; set; }
-
-        public double? Amount { get; set; }
     }
 }
