+++ ShipEngine/Models/Dto/ValidateAddresses/Enums.cs
@@ -2,17 +2,6 @@ using System.Runtime.Serialization;
 
 namespace ShipEngineSDK.ValidateAddresses
 {
-    public enum AddressResidentialIndicator
-    {
-        [EnumMember(Value = "unknown")]
-        Unknown,
-
-        [EnumMember(Value = "yes")]
-        Yes,
-
-        [EnumMember(Value = "no")]
-        No,
-    }
 
     public enum AddressValidationResult
     {
